make coverage happy by extracting unit-testable code
