+++ src/main/python/afp_cli/cli.py
@@ -25,12 +25,13 @@ import socket
 import subprocess
 import sys
 import tempfile
-from datetime import datetime
 
 import os
 import yamlreader
 from docopt import docopt
 from afp_cli import AWSFederationClientCmd, aws_credentials_file
+import afp_cli.cli_functions as cli
+
 
 CFGDIR = '/etc/afp-cli'
 DEBUG = False
@@ -113,7 +114,7 @@ def start_subshell(aws_credentials, role, account):
     print("Press CTRL+D to exit.")
     rc_script = tempfile.NamedTemporaryFile(mode='w')
     rc_script.write(RC_SCRIPT_TEMPLATE.format(role=role, account=account))
-    rc_script.write(format_aws_credentials(aws_credentials, prefix='export '))
+    rc_script.write(cli.format_aws_credentials(aws_credentials, prefix='export '))
     rc_script.flush()
     subprocess.call(
         ["bash", "--rcfile", rc_script.name],
@@ -124,7 +125,7 @@ def start_subshell(aws_credentials, role, account):
 def start_subcmd(aws_credentials, role, account):
     batch_file = tempfile.NamedTemporaryFile(suffix=".bat", delete=False)
     batch_file.write(BATCH_FILE_TEMPLATE.format(role=role, account=account))
-    batch_file.write(format_aws_credentials(aws_credentials, prefix='set '))
+    batch_file.write(cli.format_aws_credentials(aws_credentials, prefix='set '))
     batch_file.flush()
     batch_file.close()
     subprocess.call(
@@ -133,32 +134,18 @@ def start_subcmd(aws_credentials, role, account):
     os.unlink(batch_file.name)
 
 
-def get_role(arguments, federation_client, account):
-    if arguments['<rolename>']:
-        return arguments['<rolename>']
-    else:
-        try:
-            accounts_and_roles = federation_client.get_account_and_role_list()
-        except Exception as exc:
-            error("Failed to get account list from AWS: %s" % exc)
-        try:
-            return sorted(accounts_and_roles[account])[0]
-        except KeyError:
-            error("%s is not a valid AWS account" % account)
-        except IndexError:
-            error("Could not find any role for account %s" % account)
-
-
-def get_valid_seconds(aws_expiration_date, utcnow=datetime.utcnow()):
+def get_first_role(federation_client, account):
     try:
-        credentials_valid_until = datetime.strptime(aws_expiration_date, "%Y-%m-%dT%H:%M:%SZ", )
-        return (credentials_valid_until - utcnow).seconds
-    except Exception:
-        default_seconds = 3600
-        msg = "Failed to parse expiration date '{0}' for AWS credentials, assuming {1} seconds.".format(
-            aws_expiration_date, default_seconds)
-        print(msg, file=sys.stderr)
-        return default_seconds
+        accounts_and_roles = federation_client.get_account_and_role_list()
+    except Exception as exc:
+        error("Failed to get account list from AWS: %s" % exc)
+
+    try:
+        return sorted(accounts_and_roles[account])[0]
+    except KeyError:
+        error("%s is not a valid AWS account" % account)
+    except IndexError:
+        error("Could not find any role for account %s" % account)
 
 
 def get_aws_credentials(federation_client, account, role):
@@ -167,21 +154,10 @@ def get_aws_credentials(federation_client, account, role):
     except Exception as exc:
         error("Failed to get credentials from AWS: %s" % exc)
 
-    aws_credentials['AWS_VALID_SECONDS'] = get_valid_seconds(aws_credentials['AWS_EXPIRATION_DATE'])
+    aws_credentials['AWS_VALID_SECONDS'] = cli.get_valid_seconds(aws_credentials['AWS_EXPIRATION_DATE'])
     aws_credentials['AWS_ACCOUNT_NAME'] = account
     aws_credentials['AWS_ASSUMED_ROLE'] = role
     return aws_credentials
-
-
-def format_aws_credentials(credentials, prefix=''):
-    """Format aws credentials with optional prefix"""
-    return os.linesep.join(["{0}{1}='{2}'".format(prefix, key, value)
-                            for (key, value) in sorted(credentials.items())])
-
-
-def format_account_and_role_list(account_and_role_list):
-    return os.linesep.join(["{0:<20} {1}".format(account, ",".join(sorted(roles)))
-                            for account, roles in sorted(account_and_role_list.items())])
 
 
 def main():
@@ -207,17 +183,17 @@ def main():
                                                password=password)
     if arguments['<accountname>']:
         account = arguments['<accountname>']
-        role = get_role(arguments, federation_client, account)
+        role = arguments['<rolename>'] or get_first_role(federation_client, account)
         aws_credentials = get_aws_credentials(federation_client, account, role)
 
         if arguments['--show']:
-            print(format_aws_credentials(aws_credentials))
+            print(cli.format_aws_credentials(aws_credentials))
 
         elif arguments['--export']:
             if os.name == "nt":
-                print(format_aws_credentials(aws_credentials, prefix='set '))
+                print(cli.format_aws_credentials(aws_credentials, prefix='set '))
             else:
-                print(format_aws_credentials(aws_credentials, prefix='export '))
+                print(cli.format_aws_credentials(aws_credentials, prefix='export '))
         elif arguments['--write']:
             aws_credentials_file.write(aws_credentials)
         else:
@@ -232,6 +208,6 @@ def main():
                 error("Failed to start subshell: %s" % exc)
     else:
         try:
-            print(format_account_and_role_list(federation_client.get_account_and_role_list()))
+            print(cli.format_account_and_role_list(federation_client.get_account_and_role_list()))
         except Exception as exc:
             error("Failed to get account list from AWS: %s" % exc)
